+++ func/getQueryStr.ts
@@ -1,4 +1,3 @@
-
 /*
   * Extract parameters
   * 提取参数
@@ -7,15 +6,15 @@
   * @returns {String or null}.
   * @example
   * way 1
-  * window.location = 'www.example.com/page?name=jason'
-  * getQueryStr('name')
+  * window.location = "www.example.com/page?name=jason"
+  * getQueryStr("name")
   * => jason
   * way 2
-  * getQueryStr('name', 'name=jason')
+  * getQueryStr("name", "name=jason")
   * => jason
 */
-const getQueryStr = (paramName, strParams) => {
-  let reg = new RegExp('(^|&)' + paramName + '=([^&]*)(&|$)', 'i');
+const getQueryStr = (paramName: string, strParams: string) => {
+  let reg = new RegExp("(^|&)" + paramName + "=([^&]*)(&|$)", "i");
   let r = strParams?strParams.match(reg):window.location.search.substring(1).match(reg);
   if (r != null) return decodeURI(r[2]);
   return null;