+++ func/isEmpty.ts
@@ -1,5 +1,4 @@
-
-import exactType from './exactType.mjs'
+import exactType from "./exactType"
 
 /*
   * Determine if it is empty
@@ -8,44 +7,44 @@ import exactType from './exactType.mjs'
   * @param {*}.
   * @returns {Boolean}.
   * @example
-  * isEmpty('empty')
+  * isEmpty("empty")
   * => false
 */
-const isEmpty = (element) => {
+const isEmpty = (element?: any) => {
   let eleType = exactType(element);
   switch (eleType) {
-    case 'error':
-      return new TypeError('Error type parameters are invalid');
-    case 'regexp':
-      return new TypeError('Regexp type parameters are invalid');
-    case 'function':
-      return new TypeError('Function type parameters are invalid');
-    case 'date':
+    case "error":
+      return new TypeError("Error type parameters are invalid");
+    case "regexp":
+      return new TypeError("Regexp type parameters are invalid");
+    case "function":
+      return new TypeError("Function type parameters are invalid");
+    case "date":
       if(element instanceof Date && isNaN(element.getTime())) {
-        return new TypeError('Function type parameters are invalid');
+        return new TypeError("Function type parameters are invalid");
       } else {
         return false;
       }
-    case 'null':
+    case "null":
       return true;
-    case 'undefined':
+    case "undefined":
       return true;
-    case 'NaN':
+    case "NaN":
       return true;
-    case 'string':
-      if(element === ''||element === 'null'||element === 'undefined'||element === '[]'||element === '{}'||element === '[{}]') {
+    case "string":
+      if(element === ""||element === "null"||element === "undefined"||element === "[]"||element === "{}"||element === "[{}]") {
         return true;
       } else {
         return false;
       }
-    case 'symbol':
+    case "symbol":
       let symbols = Object.getOwnPropertySymbols(element);
       if(symbols.length === 0) {
         return true;
       } else {
         return false;
       }
-    case 'object':
+    case "object":
       let objectSymbols = Object.getOwnPropertySymbols(element);
       if(objectSymbols.length > 0) {
         return false;
@@ -55,7 +54,7 @@ const isEmpty = (element) => {
         }
         return true;
       }
-    case 'array':
+    case "array":
       if(element.length === 0) {
         return true;
       } else {